create algorithms modulecreate algorithms module
add abstract base class with dimensionality asserts and empty param
default choices in __call__ routine; algorithms implement the _run
abstractmethod and assume they will receive not-None inputs (as defaults
given by __call__). NOTE: still not sure if we need to have the
Algorithm iface, seems like we only have it to discern between our algo
versions and the baseline--is it helpful for communication?